Release potential fix for single error seeming to kill all goroutines
Realized there might be resource contention in other parts of the goroutine. Finally traced it back to `pool.Begin()` starting a new transaction but that doesn't get released when there's an error. Adding a `defer rollback()` seems to fix the issue.

Was able to plausibly replicate the error by making the HTTP part error w/ 100% certainty and seeing whether all scraping stopped after some point. It did! After adding the rollback, it seems to resume.